Sync functions deprecated since 3.24
