+++ gtk/gtk_deprecated_since_3_16.go
@@ -22,6 +22,14 @@ func (v *Widget) OverrideColor(state StateFlags, color *gdk.RGBA) {
 	C.gtk_widget_override_color(v.native(), C.GtkStateFlags(state), cColor)
 }
 
+func (v *Widget) OverrideBackgroundColor(state StateFlags, color *gdk.RGBA) {
+	var cColor *C.GdkRGBA
+	if color != nil {
+		cColor = (*C.GdkRGBA)(unsafe.Pointer(color.Native()))
+	}
+	C.gtk_widget_override_background_color(v.native(), C.GtkStateFlags(state), &rgba)
+}
+
 // OverrideFont is a wrapper around gtk_widget_override_font().
 func (v *Widget) OverrideFont(description string) {
 	cstr := C.CString(description)
